+++ src/api/index.ts
@@ -9,4 +9,11 @@ export { resize, resizeSync } from "./resize";
 export { toJpeg, toJpegSync, toPng, toPngSync, toWebp, toWebpSync } from "./encode";
 export { transform, transformSync } from "./transform";
 export { blurhash, blurhashSync } from "./blurhash";
+export {
+  thumbhash,
+  thumbhashSync,
+  thumbhashToRgba,
+  thumbhashToRgbaSync,
+  thumbhashToDataUrl,
+} from "./thumbhash";
 export { writeExif, writeExifSync, stripExif, stripExifSync } from "./exif";